move communities/progress to deployments endpoint
