+++ server/docs/api.md
@@ -14,6 +14,11 @@ Fuse studio API backend
 	- [Fetch community progress by id](#Fetch-community-progress-by-id)
 	- [Fetch community progress by community address](#Fetch-community-progress-by-community-address)
 	
+- [Deployment](#Deployment)
+	- [Start deployment process](#Start-deployment-process)
+	- [Fetch deployment progress by id](#Fetch-deployment-progress-by-id)
+	- [Fetch deployment progress by community address](#Fetch-deployment-progress-by-community-address)
+	
 - [Entity](#Entity)
 	- [Fetch entities by community address](#Fetch-entities-by-community-address)
 	- [Fetch entity by community address and account](#Fetch-entity-by-community-address-and-account)
@@ -126,6 +131,71 @@ GET /communities/progress/:id
 
 ```
 GET /communities/progress
+```
+
+### Parameter Parameters
+| Name     | Type       | Description                           |
+|:---------|:-----------|:--------------------------------------|
+| communityAddress | `String` | <p>Community address</p> |
+
+
+### Success 200
+| Name     | Type       | Description                           |
+|:---------|:-----------|:--------------------------------------|
+| _id | `String` | <p>Id of the deploy progress</p> |
+| steps | `Object` | <p>Steps with info about each step</p> |
+| communityAddress | `String` | <p>Community address if deploy process is completed</p> |
+| done | `Boolean` | <p>Done status if the deploy process is completed</p> |
+# <a name='Deployment'></a> Deployment
+
+## <a name='Start-deployment-process'></a> Start deployment process
+[Back to top](#top)
+
+
+
+```
+POST /communities/deploy
+```
+
+### Parameter Parameters
+| Name     | Type       | Description                           |
+|:---------|:-----------|:--------------------------------------|
+| steps | `Object` | <p>The steps (with arguments) to deploy</p> |
+
+
+### Success 200
+| Name     | Type       | Description                           |
+|:---------|:-----------|:--------------------------------------|
+| steps | `Object` | <p>Steps to deploy</p> |
+## <a name='Fetch-deployment-progress-by-id'></a> Fetch deployment progress by id
+[Back to top](#top)
+
+
+
+```
+GET /deployments/progress/:id
+```
+
+### Parameter Parameters
+| Name     | Type       | Description                           |
+|:---------|:-----------|:--------------------------------------|
+| id | `String` | <p>Auto Generated by the database</p> |
+
+
+### Success 200
+| Name     | Type       | Description                           |
+|:---------|:-----------|:--------------------------------------|
+| _id | `String` | <p>Id of the deploy progress</p> |
+| steps | `Object` | <p>Steps with info about each step</p> |
+| communityAddress | `String` | <p>Community address if deploy process is completed</p> |
+| done | `Boolean` | <p>Done status if the deploy process is completed</p> |
+## <a name='Fetch-deployment-progress-by-community-address'></a> Fetch deployment progress by community address
+[Back to top](#top)
+
+
+
+```
+GET /deployments
 ```
 
 ### Parameter Parameters
