+++ server/nodemon.json
@@ -1,0 +1,5 @@
+{
+  "ignore": [
+    "docs"
+  ]
+}
