+++ server/src/routes/api/index.js
@@ -11,6 +11,7 @@ router.use('/businesses', require('./businesses'))
 router.use('/business/list', require('./businessList'))
 router.use('/bridges', require('./bridges'))
 router.use('/communities', require('./communities'))
+router.use('/deployments', require('./deployments'))
 router.use('/entities', require('./entities'))
 
 router.use(function (err, req, res, next) {
